requirements.txt
streamlit>=1.32
openai>=1.35.0
pymupdf>=1.23.0
requests>=2.32
beautifulsoup4>=4.12
anthropic>=0.34.0
google-generativeai>=0.7.2